requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
google-genai
mangum
